requirements.txt
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.70.0
docker==7.1.0
fastapi==0.120.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
mysql==0.0.3
mysqlclient==2.2.7
numpy==1.26.4
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
psutil==6.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.14.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.23.0
boto3
mysql.connector
ruff
black 
pytest
pytest-cov
httpx